Fixed pollen bug and Joe Biden feature  addition
Adjusted the pollen reporting to account for Atlanta's non-reporting of pollen counts outside of the pollen season. Responds with a message that explains why there is no pollen count on the weekends. Added a Joe Biden feature (apolitical feature)  similar to the Tom Brady feature per the request of the server members.